i3s example: refactoring
