+++ examples/experimental/i3s-17-and-debug/components/control-panel.js
