+++ examples/experimental/i3s-17-and-debug/luma/gltf-material-parser.js
