+++ examples/experimental/i3s-17-and-debug/mesh-layer/simple-mesh-layer-fragment.glsl.js
