+++ examples/experimental/i3s-17-and-debug/mesh-layer/simple-mesh-layer-vertex.glsl.js
