+++ examples/experimental/i3s-17-and-debug/package.json
@@ -3,6 +3,7 @@
   "scripts": {
     "start": "webpack-dev-server --progress --hot --open",
     "start-local": "webpack-dev-server --env.local --env.esnext --progress --hot --open",
+    "start-local-debug": "webpack-dev-server --config webpack-debug.config.js --env.local --env.esnext --progress --hot --open",
     "start-local-deck": "webpack-dev-server --env.local --env.esnext --env.deck --progress --hot --open",
     "start-local-luma": "webpack-dev-server --env.local --env.esnext --env.luma --progress --hot --open",
     "start-local-math": "webpack-dev-server --env.local --env.esnext --env.math --progress --hot --open",
@@ -14,14 +15,19 @@
     "@deck.gl/layers": "8.4.6",
     "@deck.gl/mesh-layers": "8.4.6",
     "@deck.gl/react": "8.4.6",
-    "@loaders.gl/i3s": "^2.3.12",
-    "@luma.gl/constants": "8.4.4",
-    "@luma.gl/core": "8.4.4",
-    "@math.gl/core": "^3.3.0",
-    "@math.gl/geospatial": "^3.3.0",
-    "@probe.gl/stats-widget": "^3.3.0",
+    "@loaders.gl/core": "2.2.0",
+    "@loaders.gl/draco": "2.2.0",
+    "@loaders.gl/gltf": "2.2.0",
+    "@loaders.gl/i3s": "2.2.0",
+    "@luma.gl/constants": "^8.4.3",
+    "@luma.gl/core": "^8.4.3",
+    "@luma.gl/shadertools": "^8.4.3",
+    "@luma.gl/webgl": "^8.4.3",
+    "@math.gl/core": "^3.4.0",
+    "@math.gl/geospatial": "^3.4.0",
     "gl-matrix": "^3.3.0",
     "math.gl": "^3.4.2",
+    "@probe.gl/stats-widget": "^3.3.0",
     "probe.gl": "^3.3.0",
     "prop-types": "^15.7.2",
     "react": "^16.8.0",
