+++ examples/experimental/i3s-17-and-debug/webpack-debug.config.js
@@ -2,7 +2,7 @@ const CONFIG = {
   mode: 'development',
 
   entry: {
-    app: './app.js'
+    app: './app-debug.js'
   },
 
   output: {
