--- examples/experimental/i3s-debug/index.html
@@ -1,29 +1,0 @@
-<!doctype html>
-<html>
-<head>
-  <meta charset='UTF-8' />
-  <title>I3S Debug</title>
-  <style>
-    body {
-      font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif;
-      font-size: 13px;
-      -webkit-font-smoothing: antialiased;
-      margin:0;
-      padding:0;
-      overflow:hidden;
-      width: 100vw;
-      height: 100vh;
-    }
-    #app {
-      height: 100%;
-    }
-  </style>
-</head>
-<body/>
-<div id="app"></div>
-<script src='./app.js'></script>
-</body>
-<script type="text/javascript">
-  App.renderToDOM(document.getElementById('app'));
-</script>
-</html>
