--- examples/experimental/i3s-picking/constants.js
@@ -1,7 +1,0 @@
-export const MAP_STYLES = {
-  'Base Map: Satellite': 'https://basemaps.cartocdn.com/gl/voyager-nolabels-gl-style/style.json',
-  'Base Map: Light': 'https://basemaps.cartocdn.com/gl/positron-nolabels-gl-style/style.json',
-  'Base Map: Dark': 'https://basemaps.cartocdn.com/gl/dark-matter-nolabels-gl-style/style.json'
-};
-
-export const INITIAL_MAP_STYLE = MAP_STYLES['Base Map: Dark'];
