--- examples/experimental/i3s-picking/package.json
@@ -1,49 +1,0 @@
-{
-  "license": "MIT",
-  "scripts": {
-    "start": "webpack-dev-server --progress --hot --open",
-    "start-local": "webpack-dev-server --env.local --env.esnext --progress --hot --open",
-    "start-local-deck": "webpack-dev-server --env.local --env.esnext --env.deck --progress --hot --open",
-    "start-local-luma": "webpack-dev-server --env.local --env.esnext --env.luma --progress --hot --open",
-    "start-local-math": "webpack-dev-server --env.local --env.esnext --env.math --progress --hot --open",
-    "generate": "node ./generate-index/index.js"
-  },
-  "dependencies": {
-    "@deck.gl/core": "8.4.6",
-    "@deck.gl/geo-layers": "8.4.6",
-    "@deck.gl/layers": "8.4.6",
-    "@deck.gl/mesh-layers": "8.4.6",
-    "@deck.gl/react": "8.4.6",
-    "@loaders.gl/core": "2.2.0",
-    "@loaders.gl/draco": "2.2.0",
-    "@loaders.gl/gltf": "2.2.0",
-    "@loaders.gl/i3s": "2.2.0",
-    "@luma.gl/constants": "^8.4.3",
-    "@luma.gl/core": "^8.4.3",
-    "@luma.gl/shadertools": "^8.4.3",
-    "@luma.gl/webgl": "^8.4.3",
-    "@math.gl/core": "^3.4.0",
-    "@math.gl/geospatial": "^3.4.0",
-    "@probe.gl/stats-widget": "^3.3.0",
-    "probe.gl": "^3.3.0",
-    "prop-types": "^15.7.2",
-    "react": "^16.8.0",
-    "react-dom": "^16.8.0",
-    "react-map-gl": "^5.0.0",
-    "styled-components": "^4.2.0"
-  },
-  "devDependencies": {
-    "@babel/cli": "^7.0.0",
-    "@babel/core": "^7.4.0",
-    "@babel/preset-env": "^7.0.0",
-    "@babel/preset-react": "^7.0.0",
-    "babel-eslint": "^10.0.1",
-    "babel-loader": "^8.0.5",
-    "babel-plugin-inline-import": "^3.0.0",
-    "babel-polyfill": "^6.26.0",
-    "eslint-plugin-react": "^7.10",
-    "webpack": "^4.39.1",
-    "webpack-cli": "^3.1.2",
-    "webpack-dev-server": "^3.1.11"
-  }
-}
